\name{yamarketr-package}
\alias{yamarketr-package}
\alias{yamarketr}
\docType{package}
\title{
\packageTitle{yamarketr}
}
\description{
\packageDescription{yamarketr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{yamarketr}
\packageIndices{yamarketr}
}
\author{
\packageAuthor{yamarketr}

Maintainer: \packageMaintainer{yamarketr}
}

\keyword{ package }
