\name{yamarketrAuth}
\alias{yamarketrAuth}
\title{Authentication in Yandex Market}
\description{У вас нет необходимости использовать данную функцию отдельно, т.к. она встроена в каждую функцию пакета yamarketr.}
\usage{
  yamarketrAuth(Login = NULL,
                NewUser = FALSE,
                TokenPath = getwd())
}

\arguments{
  \item{Login}{Логин яндекса, под которым есть доступ к нужным магазинам. В этот вектор необходимо указывать логин в случае если необходимо подключаться к разным аккаунтам. В противном случае при каждом запросе к новому аккаунту - токен будет перезатираться. При указании логина в рабочей директории будет создан отдельный файл под каждый логин, в котором будут хранится нужные для работы учётные данные.}
  \item{NewUser}{логическое выражаение TRUE или FALSE, признак того, что у пользователя обязательно нужно запросить разрешение на доступ к аккаунту (даже если пользователь уже разрешил доступ данному приложению).}
  \item{TokenPath}{Путь к папке в которой хранятся все файлы с учётными данными.}
}

\details{
Вы можете в любой момент пройти аутентификацию в API Яндекс Маркет, однако это не обязательно, т.к. при запуске любой функции пакета, которая требует аутентификации будет запущен поиск файла с токеном доступа. Если файл не будет найден, то автоматически будет запущен процесс авторизации и запроса токена.

Перед выполнением аутентификации рекомендую заранее создать папку для сохранения токенов и далее объявлять её рабочей директорией перед запуском функций пакета.

Подробную инструкцию вы можете найти \href{https://github.com/mrykin/yamarketr#авторизация-в-api-яндекс-маркет}{здесь}.

}

\value{Лист и файл RData с данными авторизации.}
\author{Pavel Mrykin}


\examples{
  \dontrun{
    # Подключаем пакет
    library(yamarketr)
    # Авторизовываемся
    Token <- yamarketrAuth(Login = "login", NewUser = TRUE)
  }
}
