\name{yamarketrGetCosts}
\alias{yamarketrGetCosts}
\title{Get costs for shops from Yandex Market}
\description{Данная функция возвращает data frame с расходами по магазинам. Указание доп. параметров позволяет задать дополнительную разбивку по периодам (день/месяц/год) и местам размещения.}
\usage{
  yamarketrGetCosts(Campaigns,
                    fromDate = format(Sys.Date()-8, "%d-%m-%Y"),
                    toDate = format(Sys.Date()-1, "%d-%m-%Y"),
                    Login = NULL,
                    TokenPath = getwd(),
                    places = 0,
                    model = 0,
                    fetchBy = "daily")
}

\arguments{
  \item{Campaigns}{data frame, полученный с помощью функции yamarketrGetCampaigns, или составленный вручную, например data.frame(id=c(1111111,2222222),domain=c("shop1.ru","shop2.ru"))}
  \item{fromDate}{начальная дата отчётного периода, по умолчанию -8 дней. Формат: ДД-ММ-ГГГ ("15-03-2018")}
  \item{toDate}{конечная дата отчётного периода, по умолчанию -1 день (вчера). Формат: ДД-ММ-ГГГ ("15-03-2018")}
  \item{Login}{Логин яндекса, под которым есть доступ к нужным магазинам. В этот вектор необходимо указывать логин в случае если необходимо подключаться к разным аккаунтам. В противном случае при каждом запросе к новому аккаунту - токен будет перезатираться. При указании логина в рабочей директории будет создан отдельный файл под каждый логин, в котором будут хранится нужные для работы учётные данные.}
  \item{TokenPath}{Путь к папке в которой хранятся все файлы с учётными данными.}
  \item{places}{Признак группировки по местам размещения:
  \itemize{
  \item 1 — группировать,
  \item 0 — не группировать.}
  По умолчанию места размещения не группируются.

  Возможные значения:
  \itemize{
  \item поиск Яндекс.Маркета
  \item карточки товаров}

  \describe{
   \item{Примечание}{Значение не выводится, если во входных данных задан параметр model = 1}
          }

  \itemize{
  \item Яндекс.Маркет, кроме карточек товаров
  \item поиск Яндекса, Яндекс.Картинки, сторонние сайты и сервисы}

  Если во входных данных указан параметр model = 1, дополнительные значения идентификатора места размещения непосредственно на карточке товара:
  \itemize{
  \item предложение по умолчанию
  \item блок «Топ-6»
  \item остальные места на карточке
  }

  \itemize{
  \item иное - выводится, когда ни одна из других групп не подходит
  }

  Если группировка по местам размещения не задана в запросе, значение параметра равно 0.
  }
  \item{model}{подробная информация о месте размещения предложения на карточке модели.
  \describe{
   \item{Ограничение}{Значение model работает, только если во входных данных задан параметр places со значением: 1.}
          }
  }
  \item{fetchBy}{признак группировки по дням, неделям, месяцам ("daily", "weekly", "monthly").}
}

\value{data frame с датой расхода, id магазина, id места размещения, количество кликов, просмотров и стоимость размещения в у.е.}
\author{Pavel Mrykin}

\seealso{
  Для загрузки списка магазинов используйте \code{\link[yamarketr:yamarketrGetCampaigns]{yamarketrGetCampaigns}}
}

\examples{
  \dontrun{
    # Подключаем пакет
    library(yamarketr)
    # Получаем список магазинов
    campaigns <- yamarketrGetCampaigns(Login = "login", NewUser = FALSE)

    # Получаем расходы по магазинам за последние 7 дней, не включая сегодняшний, по дням, без разбивки по местам размещения
    costs7daysAgo <- yamarketrGetCosts(campaigns, Login = "login")

    # Получаем расходы по магазинам с 1 января до 30 сентября, по месяцам, с разбивкой по местам размещения
    costsPlaces <- yamarketrGetCosts(campaigns,
                                      Login = "login",
                                      fromDate = "01-01-2018",
                                      toDate = "30-09-2018",
                                      places = 1,
                                      model = 1,
                                      fetchBy = "monthly")
  }
}
