\name{yamarketr-package}
\alias{yamarketr-package}
\alias{yamarketr}
\docType{package}
\title{
\packageTitle{yamarketr}
}
\description{
\packageDescription{yamarketr}

Подробную инструкцию вы можете найти \href{https://github.com/mrykin/yamarketr}{здесь}
}

\author{
\packageAuthor{yamarketr}

Maintainer: \packageMaintainer{yamarketr}
}

\keyword{ package }
